temp: Allow lockfile update to regenerate yarn.lock with new dependencies
Temporarily change from `yarn install --immutable` to `yarn install` to allow the lockfile to be updated with the newly added peer dependencies (prop-types, reflect-metadata, typeorm, webpack).

After successful deployment and lockfile regeneration, this should be changed back to `yarn install --immutable` for security.

This is a one-time fix to resolve the package.json/yarn.lock mismatch.